+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailRoute.kt
@@ -1,0 +1,52 @@
+package woowacourse.shopping.ui.productdetail
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.LocalLifecycleOwner
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.repeatOnLifecycle
+import androidx.lifecycle.viewmodel.compose.viewModel
+
+@Composable
+fun ProductDetailRoute(
+    onCloseClick: () -> Unit,
+    onLastViewedProductClick: (Long) -> Unit,
+) {
+    val viewModel: ProductDetailViewModel =
+        viewModel(
+            factory = ProductDetailViewModelFactory(),
+        )
+    val snackbarHostState = remember { SnackbarHostState() }
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val lifecycleOwner = LocalLifecycleOwner.current
+
+    LaunchedEffect(viewModel.snackbarEvent, lifecycleOwner) {
+        lifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+            viewModel.snackbarEvent.collect { message ->
+                snackbarHostState.showSnackbar(message)
+            }
+        }
+    }
+
+    Scaffold(
+        snackbarHost = { SnackbarHost(snackbarHostState) },
+    ) { innerPadding ->
+        ProductDetailScreen(
+            uiState = uiState,
+            modifier = Modifier.padding(innerPadding),
+            onCloseClick = onCloseClick,
+            onLastViewedProductClick = onLastViewedProductClick,
+            onAddToCart = viewModel::addToCart,
+            onIncreaseQuantity = viewModel::increaseQuantity,
+            onDecreaseQuantity = viewModel::decreaseQuantity,
+        )
+    }
+}
